refactor(frontend): extract DashboardSuggestionCard (649 -> 434 lines)
